Move on the Grid's slides and rows using the wheel
Let the user move on the Grid's slides and rows using the mouse wheel and the
touchpad inside a GridItem.

A major difficulty to make this behaviour reliable was to prevent the Grid to
scroll too many times. This generaly happen when the user scrolls for too long,
causing the 'wheel' event to be fired too many times. It's particularly obvious
using the touchpad, as wheels events are fired for a very long time compared to
the scroll using the mouse wheel.

Three mecanisms were implemented to prevent the Grid from scrolling too many
times. The first one prevent all events to be fired during a certain amount of
time, whereas the other two help to filter those events.

The first mecanism is scroll toggling. The scroll is deactivated for a certain
amount of time after scrolling into the Grid. A reliable time seems to be 1000ms.
It's the delay used inside the FullPage.js library, so it seems to be reliable.
It must not be too long too, as the user can't scroll during this time.

Second, event throttling. Searching on the Internet, it's one of the first
solutions that is proposed to prevent events from firing too many times. To
sum it up briefly, it consists in preventing a given function to be called more
than once in a certain amount of time. A solution that comes along throttling is
debouncing, but the latter doesn't fit our purpose because the function is
called AFTER all events have been fired within a certain range one after the
other. That's not what we want, as we need to respond immediatly to the event,
and THEN to prevent undesirable events afterwards.
To use throttling, it's generaly advised to use the lodash library. I tried it
but didn't keep it, as I prefered to avoid a dependancy to add add a new feature
I could implement with only two lines of code.

Last, wheel's delta filtering. As the user scroll either using the mouse or the
touchpad, several wheel events are fired, each one having a different delta
(which represents the scroll amount). By responding only to deltas above certain
values, it helps to filter wheel events and prevent the subsequent or "residual"
events fired after the "main" scroll.
It's worth noting that delta values can change dramatically depending on the
browser. For example, as I tested, deltas on Firefox were double and rarely
exceeded 2 or 3, while on Chrome they could easily go above 90.

Each of this solution is not sufficient by itself, so we have to combine them
all to make it work. Also, I only tested this solution on my setup, so it may be
necessary to tweak the settings depending on the harware (mouse, touchpad) and
the software (OS, mouse and touchpad settings, browser, etc.). For now, I only
tested it on Windows 10 in both Firefox and Chrome.