WIP/fix/TheGrid: Fix scrolling row when GridItem not scrolled
When scrolling on a GridItem, it could immediatly move on the next row, even if
the GridItem is not scrolled (either at top or at bottom).

That is because the wheel event's target is inaccurate : it's not the GridItem 
itself, on which we check the scroll position, but another element. The scrolled
condition is not accurate, and it can lead to changing row unintentionnaly.

To fix that, we check the current GridElement directly.