WIP/GridItem: Improve moveRow ensurer
Found a new way to prevent unwanted row moves.

We no longer check the wheel delta, because:
- they are too extremes in the low and the high and
- depend on wether they are triggered using mouse wheel or touchpad,
So they are not reliable for any condition to check.

Instead, we use some sort of throttling/debounce after a GridItem has reached
either its top or its bottom. That is, when it has been scrolled, wheel events
are ignored for a certain time, repeatedly.
In other words, the move action is triggered for wheel events happening after
a minimum time a GridItem has reached either top or bottom.