fix/TheGrid: Correctly switch back GridItems when moving column
The expected behavior is that, when changing column, if the current grid's item
and the destination are on different rows, the current item is moved on the
destination's row before moving, (then we move immediatly to it, so that the
view doesn't change at all) and after the animation, it is put back at its
original position.

But there was a problem where the variable determining the column in-between,
so positions of the grid elements of the new destination columns are switched
(and it becomes a mess in both columns).

So we use a temporary variable to switch back items' positions correctly.