fix/TheGrid: Switch back GridItems even when scroll is cancelled
When clicking fast on scroll-borders, we break the first scroll animation to
start a new one. But then, the first callback is left over, and items previously
moved are not put back at their original place.

So we ensure that no matter the scroll animation is completed or cancelled, a
common callback is called to ensure that GridItems are put back.