TheGrid: Improve switching GridItems when changing column
Problems :
- Mixing onDone and onEnding callbacks is complicated
- Moving fast makes things break or behave oddly

When changing column, we now do something simpler that cover every issue :
- Before moving, on the current column, switch the current item with the item
at row 1
- After moving, on the destination column, switch back the (new) current item
with the item originally on row 1

So, before moving, all current GridItems on all different columns are on the
first row, so when scrolling columns everything will behave smoothly and
naturally.
And after moving, the current GridItem is replaced at its original position, so
that it will behave normally when scrolling rows.

Advantages :
- It modifies only on two elements at a time...
- ... elements that we know every time : item at row 1, and the current one.
- It doesn't interfere in other methods like changing row.
- It's absolute, and work in any situation.